Revert "check through regular expressions for adding arguments to our format"
