--- calendar-admin/add_interval.py
@@ -1,4 +1,0 @@
-def add_interval(namespace):
-    """добавление интервала"""
-    print("add interva {} to {}".format(
-        namespace.start, namespace.end))
