--- calendar-admin/delete_day.py
@@ -1,4 +1,0 @@
-
-def delete_day(namespace):
-    """удаляем день"""
-    print("delete {} day".format(namespace.date))
