--- calendar-admin/delete_interval.py
@@ -1,5 +1,0 @@
-
-def delete_interval(namespace):
-    """удаляем интервал"""
-    print("delete, interval {} to {}".format(
-        namespace.start, namespace.end))
