--- calendar-admin/main.py
@@ -1,5 +1,0 @@
-#!/Usr/bin/env python
-import comand_line_parser as par
-
-
-par.check_func()
